fix: review suggestions
