+++ apps/web/app/(app)/environments/[environmentId]/project/layout.tsx
@@ -1,3 +1,4 @@
-import { ProjectSettingsLayout } from "@/modules/projects/settings/layout";
+import { ProjectSettingsLayout, metadata } from "@/modules/projects/settings/layout";
 
+export { metadata };
 export default ProjectSettingsLayout;
