+++ apps/web/app/(auth)/auth/login/page.tsx
@@ -1,3 +1,4 @@
-import { LoginPage } from "@/modules/auth/login/page";
+import { LoginPage, metadata } from "@/modules/auth/login/page";
 
+export { metadata };
 export default LoginPage;
