+++ apps/web/modules/ee/role-management/lib/invite.ts
@@ -1,14 +1,10 @@
 import { inviteCache } from "@/lib/cache/invite";
-import { type TInviteUpdateInput, ZInviteUpdateInput } from "@/modules/ee/role-management/types/invites";
+import { type TInviteUpdateInput } from "@/modules/ee/role-management/types/invites";
 import { Prisma } from "@prisma/client";
 import { prisma } from "@formbricks/database";
-import { validateInputs } from "@formbricks/lib/utils/validate";
-import { ZString } from "@formbricks/types/common";
 import { ResourceNotFoundError } from "@formbricks/types/errors";
 
 export const updateInvite = async (inviteId: string, data: TInviteUpdateInput): Promise<boolean> => {
-  validateInputs([inviteId, ZString], [data, ZInviteUpdateInput]);
-
   try {
     const invite = await prisma.invite.update({
       where: { id: inviteId },
