+++ apps/web/modules/organization/settings/teams/components/edit-memberships/organization-actions.tsx
@@ -84,7 +84,7 @@ export const OrganizationActions = ({
             email: email.toLowerCase(),
             name,
             role,
-            teamIds: teamIds,
+            teamIds,
           });
           return {
             email,
