+++ apps/web/modules/setup/organization/[organizationId]/invite/components/invite-members.tsx
@@ -1,11 +1,11 @@
 "use client";
 
-import { inviteOrganizationMemberAction } from "@/app/setup/organization/[organizationId]/invite/actions";
+import { getFormattedErrorMessage } from "@/lib/utils/helper";
+import { inviteOrganizationMemberAction } from "@/modules/setup/organization/[organizationId]/invite/actions";
 import {
   type TInviteMembersFormSchema,
   ZInviteMembersFormSchema,
-} from "@/app/setup/organization/[organizationId]/invite/types/invites";
-import { getFormattedErrorMessage } from "@/lib/utils/helper";
+} from "@/modules/setup/organization/[organizationId]/invite/types/invites";
 import { Alert, AlertDescription, AlertTitle } from "@/modules/ui/components/alert";
 import { Button } from "@/modules/ui/components/button";
 import { FormControl, FormError, FormField, FormItem, FormProvider } from "@/modules/ui/components/form";
