+++ apps/web/modules/setup/organization/[organizationId]/invite/types/invites.ts
@@ -3,9 +3,8 @@ import { ZInvite } from "@formbricks/database/zod/invites";
 import { ZUserName } from "@formbricks/types/user";
 
 export const ZInvitee = ZInvite.pick({
+  name: true,
   email: true,
-  role: true,
-  teamIds: true,
 }).extend({
   name: ZUserName,
 });
